main.go
package main

import (
	"encoding/json"
	"log"
	"net/http"

	"github.com/gorilla/mux"
	"github.com/move-ton/ever-client-go/domain"
	goton "github.com/move-ton/ever-client-go"
)

// Message represents a message to be sent to the blockchain
type Message struct {
	Content string `json:"content"`
}

// Transaction represents a blockchain transaction
type Transaction struct {
	ID     string `json:"id"`
	Amount int    `json:"amount"`
}

// AccountState represents the state of an account on the blockchain
type AccountState struct {
	Address string `json:"address"`
	Balance int    `json:"balance"`
}

// BlockchainData represents the data related to the blockchain
type BlockchainData struct {
	Blocks       []Block        `json:"blocks"`
	Transactions []Transaction  `json:"transactions"`
	Messages     []BlockchainMessage `json:"messages"`
}

// Block represents a block in the blockchain
type Block struct {
	ID   string `json:"id"`
	Hash string `json:"hash"`
}

// BlockchainMessage represents a message in the blockchain
type BlockchainMessage struct {
	ID      string `json:"id"`
	Content string `json:"content"`
}

// CreateMessageHandler handles the creation and sending of messages to the blockchain
func CreateMessageHandler(w http.ResponseWriter, r *http.Request) {
	var message Message
	err := json.NewDecoder(r.Body).Decode(&message)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		return
	}

	// Process and send the message to the Venom blockchain using TVM
	err = ProcessAndSendToBlockchain(message)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(message)
}

// ProcessAndSendToBlockchain processes and sends the message to the Venom blockchain using TVM
func ProcessAndSendToBlockchain(message Message) error {
	// Initialize the Venom blockchain client
	ever, err := goton.NewEver("", domain.GetDevNetBaseUrls(), "")
	if err != nil {
		return err
	}

	defer ever.Client.Destroy()

	// Convert the message content to bytes
	messageData := []byte(message.Content)

	// Send the message to the blockchain using the Venom SDK
	_, err = ever.Client.SendMessage(messageData)
	if err != nil {
		return err
	}

	return nil
}

// GetMessageHandler retrieves a message from the blockchain
func GetMessageHandler(w http.ResponseWriter, r *http.Request) {
	// Retrieve the message from the blockchain
	// You would need to implement this logic using the Venom blockchain and TVM

	message := Message{
		Content: "Hello, world!",
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(message)
}

// GetAccountStateHandler retrieves the state of an account on the blockchain
func GetAccountStateHandler(w http.ResponseWriter, r *http.Request) {
	// Retrieve the account state from the blockchain
	// You would need to implement this logic using the Venom blockchain and TVM

	accountState := AccountState{
		Address: "0x123abc",
		Balance: 100,
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(accountState)
}

// QueryBlockchainDataHandler queries various data from the blockchain
func QueryBlockchainDataHandler(w http.ResponseWriter, r *http.Request) {
	// Query blockchain data (blocks, transactions, messages)
	// You would need to implement this logic using the Venom blockchain and TVM

	blockchainData := BlockchainData{
		Blocks: []Block{
			{ID: "block1", Hash: "hash1"},
			{ID: "block2", Hash: "hash2"},
		},
		Transactions: []Transaction{
			{ID: "tx1", Amount: 10},
			{ID: "tx2", Amount: 20},
		},
		Messages: []BlockchainMessage{
			{ID: "msg1", Content: "message1"},
			{ID: "msg2", Content: "message2"},
		},
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(blockchainData)
}

// SubscribeToEventsHandler subscribes to blockchain events and updates
func SubscribeToEventsHandler(w http.ResponseWriter, r *http.Request) {
	// Subscribe to events and any other blockchain updates
	// You would need to implement this logic using the Venom blockchain and TVM

	// Placeholder response
	response := map[string]string{
		"message": "Subscribed to events",
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}

// SignDataHandler signs the provided data and returns the signature
func SignDataHandler(w http.ResponseWriter, r *http.Request) {
	// Sign data/check signature
	// You would need to implement this logic using the Venom blockchain and TVM

	// Placeholder response
	response := map[string]string{
		"signature": "abcd1234",
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}

// CalculateHashHandler calculates the hash of the provided data
func CalculateHashHandler(w http.ResponseWriter, r *http.Request) {
	// Calculate hashes (sha256, sha512)
	// You would need to implement this logic using the Venom blockchain and TVM

	// Placeholder response
	response := map[string]string{
		"hash": "hash123",
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}

// EncryptDataHandler encrypts the provided data
func EncryptDataHandler(w http.ResponseWriter, r *http.Request) {
	// Encrypt/decrypt data
	// You would need to implement this logic using the Venom blockchain and TVM

	// Placeholder response
	response := map[string]string{
		"encryptedData": "encrypted123",
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}

// ValidateAddressHandler validates the provided address
func ValidateAddressHandler(w http.ResponseWriter, r *http.Request) {
	// Validate addresses
	// You would need to implement this logic using the Venom blockchain and TVM

	// Placeholder response
	response := map[string]bool{
		"isValid": true,
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}

// WorkWithBlockchainNativeTypesHandler handles operations with blockchain native types
func WorkWithBlockchainNativeTypesHandler(w http.ResponseWriter, r *http.Request) {
	// Work with blockchain native types (BOCs)
	// You would need to implement this logic using the Venom blockchain and TVM

	// Placeholder response
	response := map[string]string{
		"result": "success",
	}

	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}

func main() {
	r := mux.NewRouter()
	r.HandleFunc("/messages", CreateMessageHandler).Methods("POST")
	r.HandleFunc("/messages/{id}", GetMessageHandler).Methods("GET")
	r.HandleFunc("/accounts/{address}", GetAccountStateHandler).Methods("GET")
	r.HandleFunc("/blockchain", QueryBlockchainDataHandler).Methods("GET")
	r.HandleFunc("/subscribe", SubscribeToEventsHandler).Methods("POST")
	r.HandleFunc("/sign", SignDataHandler).Methods("POST")
	r.HandleFunc("/hash", CalculateHashHandler).Methods("POST")
	r.HandleFunc("/encrypt", EncryptDataHandler).Methods("POST")
	r.HandleFunc("/validate", ValidateAddressHandler).Methods("POST")
	r.HandleFunc("/bocs", WorkWithBlockchainNativeTypesHandler).Methods("POST")

	log.Fatal(http.ListenAndServe(":8080", r))
}
